server/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
psycopg==3.1.18
psycopg-binary==3.1.18
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
# TODO - look into using Poetry (the library)